+++ frontend/src/ContextAPI/PostsContext.tsx
@@ -1,6 +1,7 @@
 import React, { createContext, useState, useContext, ReactNode } from 'react';
 import { AccountContext } from './AccountContext';
 import { postService } from '../apis/postsService';
+import { AxiosResponse } from 'axios';
 
 export interface PostInfo {
   id?: number;
@@ -12,16 +13,16 @@ export interface PostInfo {
 
 interface PostsContextType {
   posts: PostInfo[];
-  getPosts: (id?: string) => Promise<void>;
-  createPost: (post: PostInfo) => Promise<boolean>;
-  deletePost: (id: number) => Promise<boolean>;
+  getPosts: (id?: string) => Promise<AxiosResponse>;
+  createPost: (post: PostInfo) => Promise<AxiosResponse>;
+  deletePost: (id: number) => Promise<AxiosResponse>;
 }
 
 const PostsContext = createContext<PostsContextType>({
   posts: [],
-  getPosts: async () => {},
-  createPost: async () => false,
-  deletePost: async () => false,
+  getPosts: async () => null,
+  createPost: async () => null,
+  deletePost: async () => null,
 });
 
 interface PostsProviderProps {
@@ -37,42 +38,49 @@ const PostsContextProvider: React.FC<PostsProviderProps> = ({ children }) => {
     try {
       const response = await postService.getPosts(id);
       setPosts(response.data);
+      return response;
     } catch (error) {
       console.error('Error fetching posts:', error);
+      throw new Error(error);
     }
   };
 
   const createPost = async (post: PostInfo) => {
+    if (!sessionJWT) {
+      throw new Error('로그인이 필요한 기능입니다.');
+    }
+
     try {
       const response = await postService.createPost(post, sessionJWT);
 
-      if (response.status === 200) {
+      if (response.status === 201) {
         setPosts([...posts, response.data]);
-        return true;
+        return response;
       }
     } catch (error) {
       console.error('Error creating post:', error);
+      return error;
     }
   };
 
   const deletePost = async (id: number) => {
+    if (!sessionJWT) {
+      throw new Error('로그인이 필요한 기능입니다.');
+    }
+
     try {
       const response = await postService.deletePost(id, sessionJWT);
 
       if (response.status === 200) {
         setPosts(posts.filter(post => post.id !== id));
-        return true;
-      } else {
-        console.error('Error deleting post:', response);
-        alert('글 삭제에 실패했습니다.');
-        return false;
+        return response;
       }
     } catch (error) {
       console.error('Error deleting post:', error);
       if (error.response.status === 403) {
-        alert('권한이 없습니다.');
+        throw new Error('권한이 없습니다.');
       } else {
-        alert('글 삭제에 실패했습니다.');
+        throw new Error('글 삭제에 실패했습니다.');
       }
     }
   };
